Add configure check X_AC_FLUX
Based on X_AC_LSF.

Use flux_open() function provided by flux-core.so to determine whether
flux is installed.

Getting actual remaining time in an allocation depends on getting a json
description of the allocation using functions in flux-core.so, and
extracting the remaining time using jansson.so.  This will be in a
separate commit.