Add support for flux
Look in the environment for FLUX_JOB_ID.  Parse it to obtain
the 64-bit unsigned integer representation and store it.

Determine whether to query the current flux instance or the parent
for the expiration time of the allocation.  Note that this
currently works by checking the environment for FLUX_KVS_NAMESPACE,
but flux will provide a more explicit mechanism in the future.
See https://github.com/flux-framework/flux-core/issues/3817
for details and status.

Fetch the expiration time and calculate remaining time based on that.

Support get_rank() by looking in the environment for FLUX_TASK_RANK.